chore(release): merge dev into main branch

+++ packages/vscode-extension/src/debug/teamsfxTaskHandler.ts
@@ -17,7 +17,10 @@ import {
   TelemetryMeasurements,
   TelemetryProperty,
 } from "../telemetry/extTelemetryEvents";
-import { Correlator, getHashedEnv, isValidProject } from "@microsoft/teamsfx-core";
+import { getHashedEnv } from "@microsoft/teamsfx-core/build/common/tools";
+import { TaskCommand } from "@microsoft/teamsfx-core/build/common/local/constants";
+import { Correlator } from "@microsoft/teamsfx-core/build/common/correlator";
+import { isValidProject } from "@microsoft/teamsfx-core/build/common/projectSettingsHelper";
 import * as path from "path";
 import {
   errorDetail,
@@ -37,6 +40,7 @@ import { VS_CODE_UI } from "../extension";
 import { localTelemetryReporter, sendDebugAllEvent } from "./localTelemetryReporter";
 import { ExtensionErrors, ExtensionSource } from "../error";
 import { performance } from "perf_hooks";
+import { LocalTunnelTaskTerminal } from "./taskTerminal/localTunnelTaskTerminal";
 
 class NpmInstallTaskInfo {
   private startTime: number;
@@ -83,6 +87,16 @@ function isNpmInstallTask(task: vscode.Task): boolean {
     return task.name.trim().toLocaleLowerCase().endsWith("npm install");
   }
 
+  return false;
+}
+
+function isTeamsFxTransparentTask(task: vscode.Task): boolean {
+  if (task.definition && task.definition.type === ProductName) {
+    const command = task.definition.command as string;
+    if ((Object.values(TaskCommand) as string[]).includes(command)) {
+      return true;
+    }
+  }
   return false;
 }
 
@@ -288,6 +302,13 @@ async function onDidEndTaskProcessHandler(event: vscode.TaskProcessEndEvent): Pr
       [TelemetryProperty.DebugServiceName]: task.name,
       [TelemetryProperty.DebugServiceExitCode]: event.exitCode + "",
     });
+  } else if (
+    task.scope !== undefined &&
+    isTeamsFxTransparentTask(task) &&
+    event.exitCode !== 0 &&
+    event.exitCode !== -1
+  ) {
+    terminateAllRunningTeamsfxTasks();
   } else if (isNpmInstallTask(task)) {
     try {
       const taskInfo = activeNpmInstallTasks.get(task.name);
@@ -470,7 +491,7 @@ async function onDidStartDebugSessionHandler(event: vscode.DebugSession): Promis
           return;
         }
 
-        await sendDebugAllEvent();
+        sendDebugAllEvent();
       }
     }
   }
@@ -479,12 +500,15 @@ async function onDidStartDebugSessionHandler(event: vscode.DebugSession): Promis
 export function terminateAllRunningTeamsfxTasks(): void {
   for (const task of allRunningTeamsfxTasks) {
     try {
-      process.kill(task[1], "SIGTERM");
+      if (task[1] > 0) {
+        process.kill(task[1], "SIGTERM");
+      }
     } catch (e) {
       // ignore and keep killing others
     }
   }
   allRunningTeamsfxTasks.clear();
+  LocalTunnelTaskTerminal.stopAll();
 }
 
 function onDidTerminateDebugSessionHandler(event: vscode.DebugSession): void {
